fix: move documentation builder to docs/md and fix yaml path handling
- Add docs/md/docs_builder.py: Centralized documentation builder
  - Processes markdown files with embedded BSL queries
  - Changes to content directory for relative file path resolution
  - Generates JSON data files for web documentation
- Remove docs/web/scripts/build_data.py: Replaced by docs_builder.py
- Update docs/web/package.json: Point build:bsl to ../md/docs_builder.py
- Fix docs/md/doc/yaml-config.md: Use relative path for yaml_example.yaml
- Fix linting issues: Remove unused variable, simplify nested if, use | syntax
- Update codespell config: Add 'notin' to ignore list (ast.NotIn is valid Python)
- Regenerate all documentation JSON files with updated builder

This completes the documentation restructure, fixing the build process
that was broken after separating content (docs/md/doc) from web app
(docs/web).
