+++ docs/web/public/bsl-data/charting.json
@@ -311,7 +311,7 @@
           }
         },
         "data": {
-          "name": "data-57b57b79a99936f4fb4fa3751e91916e"
+          "name": "data-bfb7107d7325a594fdc8cf99b8cb9a41"
         },
         "mark": {
           "type": "line"
@@ -345,7 +345,15 @@
         "width": 700,
         "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
         "datasets": {
-          "data-57b57b79a99936f4fb4fa3751e91916e": [
+          "data-bfb7107d7325a594fdc8cf99b8cb9a41": [
+            {
+              "date": "2024-01-01T00:00:00",
+              "flight_count": 145
+            },
+            {
+              "date": "2024-01-07T00:00:00",
+              "flight_count": 143
+            },
             {
               "date": "2024-01-06T00:00:00",
               "flight_count": 161
@@ -353,10 +361,6 @@
             {
               "date": "2024-01-04T00:00:00",
               "flight_count": 139
-            },
-            {
-              "date": "2024-01-01T00:00:00",
-              "flight_count": 145
             },
             {
               "date": "2024-01-02T00:00:00",
@@ -369,10 +373,6 @@
             {
               "date": "2024-01-05T00:00:00",
               "flight_count": 156
-            },
-            {
-              "date": "2024-01-07T00:00:00",
-              "flight_count": 143
             }
           ]
         }
@@ -390,16 +390,20 @@
         ],
         "data": [
           [
-            "2024-01-01",
-            145
+            "2024-01-07",
+            143
+          ],
+          [
+            "2024-01-04",
+            139
           ],
           [
             "2024-01-06",
             161
           ],
           [
-            "2024-01-07",
-            143
+            "2024-01-01",
+            145
           ],
           [
             "2024-01-02",
@@ -412,10 +416,6 @@
           [
             "2024-01-05",
             156
-          ],
-          [
-            "2024-01-04",
-            139
           ]
         ]
       },
@@ -429,7 +429,7 @@
             }
           },
           "data": {
-            "name": "data-a9ad2e2e073614169b0072a9d4cf5009"
+            "name": "data-ae0e4064b6838a7e781b80ba7a2a6667"
           },
           "mark": {
             "type": "line"
@@ -463,7 +463,19 @@
           "width": 700,
           "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
           "datasets": {
-            "data-a9ad2e2e073614169b0072a9d4cf5009": [
+            "data-ae0e4064b6838a7e781b80ba7a2a6667": [
+              {
+                "date": "2024-01-06T00:00:00",
+                "flight_count": 161
+              },
+              {
+                "date": "2024-01-01T00:00:00",
+                "flight_count": 145
+              },
+              {
+                "date": "2024-01-07T00:00:00",
+                "flight_count": 143
+              },
               {
                 "date": "2024-01-02T00:00:00",
                 "flight_count": 152
@@ -479,18 +491,6 @@
               {
                 "date": "2024-01-04T00:00:00",
                 "flight_count": 139
-              },
-              {
-                "date": "2024-01-06T00:00:00",
-                "flight_count": 161
-              },
-              {
-                "date": "2024-01-01T00:00:00",
-                "flight_count": 145
-              },
-              {
-                "date": "2024-01-07T00:00:00",
-                "flight_count": 143
               }
             ]
           }
@@ -506,7 +506,7 @@
           }
         },
         "data": {
-          "name": "data-bc6c0c48f000287c36baeeb747eb8c67"
+          "name": "data-bd539b775c5bc5f4f5724f6ec18e7e90"
         },
         "mark": {
           "type": "rect"
@@ -545,7 +545,17 @@
         "width": 700,
         "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
         "datasets": {
-          "data-bc6c0c48f000287c36baeeb747eb8c67": [
+          "data-bd539b775c5bc5f4f5724f6ec18e7e90": [
+            {
+              "origin": "SFO",
+              "dest": "JFK",
+              "flight_count": 31
+            },
+            {
+              "origin": "JFK",
+              "dest": "LAX",
+              "flight_count": 45
+            },
             {
               "origin": "LAX",
               "dest": "SFO",
@@ -565,16 +575,6 @@
               "origin": "LAX",
               "dest": "JFK",
               "flight_count": 43
-            },
-            {
-              "origin": "JFK",
-              "dest": "LAX",
-              "flight_count": 45
-            },
-            {
-              "origin": "SFO",
-              "dest": "JFK",
-              "flight_count": 31
             }
           ]
         }
@@ -590,7 +590,7 @@
           }
         },
         "data": {
-          "name": "data-b871f1454fe39c6ec6f7cefaa7083022"
+          "name": "data-7b03f826e12a6baefce267389b20b919"
         },
         "mark": {
           "type": "bar"
@@ -643,26 +643,26 @@
         "width": 700,
         "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
         "datasets": {
-          "data-b871f1454fe39c6ec6f7cefaa7083022": [
-            {
-              "origin": "SFO",
-              "flight_count": 89,
-              "avg_distance": 1200.0
-            },
-            {
-              "origin": "ATL",
-              "flight_count": 145,
-              "avg_distance": 1650.0
-            },
+          "data-7b03f826e12a6baefce267389b20b919": [
             {
               "origin": "DEN",
               "flight_count": 78,
               "avg_distance": 900.0
             },
             {
+              "origin": "LAX",
+              "flight_count": 135,
+              "avg_distance": 1850.0
+            },
+            {
               "origin": "DFW",
               "flight_count": 98,
               "avg_distance": 1100.0
+            },
+            {
+              "origin": "ORD",
+              "flight_count": 112,
+              "avg_distance": 950.0
             },
             {
               "origin": "JFK",
