+++ docs/web/public/bsl-data/comparison.json
