+++ docs/web/public/bsl-data/example.json
@@ -3,7 +3,7 @@
   "queries": {
     "revenue_by_customer": {
       "code": "result = orders_st.group_by(\"customer\").aggregate(\n    \"total_orders\",\n    \"total_revenue\",\n    \"avg_order_value\"\n)",
-      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"customer\",\n    COUNT(*) AS \"total_orders\",\n    SUM(\"t1\".\"amount\") AS \"total_revenue\",\n    AVG(\"t1\".\"amount\") AS \"avg_order_value\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_azthjennpzct7juvrsqfyt6a7e\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
+      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"customer\",\n    COUNT(*) AS \"total_orders\",\n    SUM(\"t1\".\"amount\") AS \"total_revenue\",\n    AVG(\"t1\".\"amount\") AS \"avg_order_value\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_dv7tvk7s65f4fjs3kudioehnca\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
       "plan": "r0 := SemanticTableOp[orders]\n  2 dimensions: customer, product\n  4 measures: total_orders, total_revenue, total_quantity, ...\n\nr1 := SemanticGroupByOp\n  source: SemanticTableOp\n  keys: ['customer']\n  inherited dimensions: 2\n  inherited measures: 4\n\nSemanticAggregateOp\n  source: SemanticGroupByOp\n  keys: ['customer']\n  aggs: [total_orders, total_revenue, avg_order_value]",
       "table": {
         "columns": [
@@ -14,10 +14,10 @@
         ],
         "data": [
           [
-            "David",
-            1,
-            100,
-            100.0
+            "Bob",
+            2,
+            250,
+            125.0
           ],
           [
             "Charlie",
@@ -26,16 +26,16 @@
             112.5
           ],
           [
+            "David",
+            1,
+            100,
+            100.0
+          ],
+          [
             "Alice",
             3,
             350,
             116.66666666666667
-          ],
-          [
-            "Bob",
-            2,
-            250,
-            125.0
           ]
         ]
       },
@@ -49,7 +49,7 @@
             }
           },
           "data": {
-            "name": "data-3bfbb7e6333f05bbae96f33cd7cab22a"
+            "name": "data-e6a2e97bb8df58131dd2d01a4f00c855"
           },
           "mark": {
             "type": "bar"
@@ -103,13 +103,7 @@
           "width": 700,
           "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
           "datasets": {
-            "data-3bfbb7e6333f05bbae96f33cd7cab22a": [
-              {
-                "customer": "David",
-                "total_orders": 1,
-                "total_revenue": 100,
-                "avg_order_value": 100.0
-              },
+            "data-e6a2e97bb8df58131dd2d01a4f00c855": [
               {
                 "customer": "Charlie",
                 "total_orders": 2,
@@ -117,16 +111,22 @@
                 "avg_order_value": 112.5
               },
               {
-                "customer": "Alice",
-                "total_orders": 3,
-                "total_revenue": 350,
-                "avg_order_value": 116.66666666666667
+                "customer": "David",
+                "total_orders": 1,
+                "total_revenue": 100,
+                "avg_order_value": 100.0
               },
               {
                 "customer": "Bob",
                 "total_orders": 2,
                 "total_revenue": 250,
                 "avg_order_value": 125.0
+              },
+              {
+                "customer": "Alice",
+                "total_orders": 3,
+                "total_revenue": 350,
+                "avg_order_value": 116.66666666666667
               }
             ]
           }
@@ -135,7 +135,7 @@
     },
     "product_performance": {
       "code": "result = orders_st.group_by(\"product\").aggregate(\n    \"total_orders\",\n    \"total_quantity\",\n    \"total_revenue\"\n)",
-      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"product\",\n    COUNT(*) AS \"total_orders\",\n    SUM(\"t1\".\"quantity\") AS \"total_quantity\",\n    SUM(\"t1\".\"amount\") AS \"total_revenue\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_azthjennpzct7juvrsqfyt6a7e\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
+      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"product\",\n    COUNT(*) AS \"total_orders\",\n    SUM(\"t1\".\"quantity\") AS \"total_quantity\",\n    SUM(\"t1\".\"amount\") AS \"total_revenue\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_dv7tvk7s65f4fjs3kudioehnca\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
       "plan": "r0 := SemanticTableOp[orders]\n  2 dimensions: customer, product\n  4 measures: total_orders, total_revenue, total_quantity, ...\n\nr1 := SemanticGroupByOp\n  source: SemanticTableOp\n  keys: ['product']\n  inherited dimensions: 2\n  inherited measures: 4\n\nSemanticAggregateOp\n  source: SemanticGroupByOp\n  keys: ['product']\n  aggs: [total_orders, total_quantity, total_revenue]",
       "table": {
         "columns": [
@@ -152,16 +152,16 @@
             75
           ],
           [
-            "Gadget",
-            3,
-            6,
-            450
-          ],
-          [
             "Widget",
             4,
             4,
             400
+          ],
+          [
+            "Gadget",
+            3,
+            6,
+            450
           ]
         ]
       },
