+++ docs/web/public/bsl-data/getting-started.json
@@ -3,7 +3,7 @@
   "queries": {
     "query_by_origin": {
       "code": "# Group flights by origin airport\nresult = flights_st.group_by(\"origin\").aggregate(\n    \"flight_count\",\n    \"total_distance\",\n    \"avg_duration\"\n)",
-      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"origin\",\n    COUNT(*) AS \"flight_count\",\n    SUM(\"t1\".\"distance\") AS \"total_distance\",\n    AVG(\"t1\".\"duration\") AS \"avg_duration\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_tahjc4mqz5hivojrsiwqjqw35u\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
+      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"origin\",\n    COUNT(*) AS \"flight_count\",\n    SUM(\"t1\".\"distance\") AS \"total_distance\",\n    AVG(\"t1\".\"duration\") AS \"avg_duration\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_sfl7eggqanetdm5hsfvizxcauq\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
       "plan": "r0 := SemanticTableOp[flights]\n  2 dimensions: origin, destination\n  3 measures: flight_count, total_distance, avg_duration\n\nr1 := SemanticGroupByOp\n  source: SemanticTableOp\n  keys: ['origin']\n  inherited dimensions: 2\n  inherited measures: 3\n\nSemanticAggregateOp\n  source: SemanticGroupByOp\n  keys: ['origin']\n  aggs: [flight_count, total_distance, avg_duration]",
       "table": {
         "columns": [
@@ -20,16 +20,16 @@
             340.0
           ],
           [
-            "LAX",
-            3,
-            3483,
-            153.33333333333334
-          ],
-          [
             "SFO",
             2,
             3249,
             212.5
+          ],
+          [
+            "LAX",
+            3,
+            3483,
+            153.33333333333334
           ]
         ]
       },
@@ -43,7 +43,7 @@
             }
           },
           "data": {
-            "name": "data-0137bc47e458ebaf211d83de5a1f535d"
+            "name": "data-6a5c6500dd6e1fd5a2687fe537305e7e"
           },
           "mark": {
             "type": "bar"
@@ -97,12 +97,12 @@
           "width": 700,
           "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
           "datasets": {
-            "data-0137bc47e458ebaf211d83de5a1f535d": [
+            "data-6a5c6500dd6e1fd5a2687fe537305e7e": [
               {
-                "origin": "NYC",
-                "flight_count": 3,
-                "total_distance": 8480,
-                "avg_duration": 340.0
+                "origin": "SFO",
+                "flight_count": 2,
+                "total_distance": 3249,
+                "avg_duration": 212.5
               },
               {
                 "origin": "LAX",
@@ -111,10 +111,10 @@
                 "avg_duration": 153.33333333333334
               },
               {
-                "origin": "SFO",
-                "flight_count": 2,
-                "total_distance": 3249,
-                "avg_duration": 212.5
+                "origin": "NYC",
+                "flight_count": 3,
+                "total_distance": 8480,
+                "avg_duration": 340.0
               }
             ]
           }
@@ -123,7 +123,7 @@
     },
     "query_by_destination": {
       "code": "# Group flights by destination airport\nresult = flights_st.group_by(\"destination\").aggregate(\n    \"flight_count\",\n    \"total_distance\"\n)",
-      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"destination\",\n    COUNT(*) AS \"flight_count\",\n    SUM(\"t1\".\"distance\") AS \"total_distance\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_tahjc4mqz5hivojrsiwqjqw35u\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
+      "sql": "SELECT\n  *\nFROM (\n  SELECT\n    \"t1\".\"destination\",\n    COUNT(*) AS \"flight_count\",\n    SUM(\"t1\".\"distance\") AS \"total_distance\"\n  FROM (\n    SELECT\n      *\n    FROM \"ibis_pandas_memtable_sfl7eggqanetdm5hsfvizxcauq\" AS \"t0\"\n  ) AS \"t1\"\n  GROUP BY\n    1\n) AS \"t2\"",
       "plan": "r0 := SemanticTableOp[flights]\n  2 dimensions: origin, destination\n  3 measures: flight_count, total_distance, avg_duration\n\nr1 := SemanticGroupByOp\n  source: SemanticTableOp\n  keys: ['destination']\n  inherited dimensions: 2\n  inherited measures: 3\n\nSemanticAggregateOp\n  source: SemanticGroupByOp\n  keys: ['destination']\n  aggs: [flight_count, total_distance]",
       "table": {
         "columns": [
@@ -159,7 +159,7 @@
             }
           },
           "data": {
-            "name": "data-054f94824247692bd2f208176250cd44"
+            "name": "data-67d2644c1dafad532d6ad5112b2e2fe3"
           },
           "mark": {
             "type": "bar"
@@ -212,11 +212,11 @@
           "width": 700,
           "$schema": "https://vega.github.io/schema/vega-lite/v5.20.1.json",
           "datasets": {
-            "data-054f94824247692bd2f208176250cd44": [
+            "data-67d2644c1dafad532d6ad5112b2e2fe3": [
               {
-                "destination": "LAX",
-                "flight_count": 3,
-                "total_distance": 5925
+                "destination": "NYC",
+                "flight_count": 2,
+                "total_distance": 5691
               },
               {
                 "destination": "SFO",
@@ -224,9 +224,9 @@
                 "total_distance": 3596
               },
               {
-                "destination": "NYC",
-                "flight_count": 2,
-                "total_distance": 5691
+                "destination": "LAX",
+                "flight_count": 3,
+                "total_distance": 5925
               }
             ]
           }
