+++ docs/web/public/pages.json
@@ -1,1 +1,18 @@
-["sessionized", "reference", "comparison", "semantic-table", "example", "mcp", "query-methods", "indexing", "bucketing", "getting-started", "compose", "percentage-total", "nested-subtotals", "yaml-config", "windowing", "charting"]
+[
+  "bucketing",
+  "charting",
+  "comparison",
+  "compose",
+  "example",
+  "getting-started",
+  "indexing",
+  "mcp",
+  "nested-subtotals",
+  "percentage-total",
+  "query-methods",
+  "reference",
+  "semantic-table",
+  "sessionized",
+  "windowing",
+  "yaml-config"
+]
